Move Cluster out of the config package
This makes it clearer that the config package is independent of it and
it is just an example of how one can use the config package.
